refactor quadrature things out of ElementUtils for clarity
